Use dependency injection for ClassLoader
Avoid CacheManager usage, as that would be needed to be
lazy loaded, as we need to instantiate our ClassLoader
instance during ext_localconf.php load.
Use Bootsrap::createCache as factory for our cache.
Adapt ClassCacheManager to use the this cache as well.

We need manual wiring for our custom cache anyway,
so we disable autowiring.

TODO:
 * Maybe refactor to be executed once Core\Bootstrap::init completes.
   Problem: There is no event/hook available.
   Possible solution: We could theoretically intercept the Core\Context
   class factory, that class is loaded for every request (frontend,backend,cli).
   Which means we could register our autoloader at that time.