Register autoloader via an early event
The event dispatcher is created prior loading of
ext_localconf, therefore an early event is now
attached in order to load the extender autoloader
at an early stage.

The event is added via a compiler pass that modifies
the event dispatcher factory in order to dispatch
the desired event on demand.